pyproject.toml
[tool.poetry]
name = "kanban-board"
version = "0.1.0"
description = ""
authors = ["Oleg Korshunov <oleg.korshunov@simbirsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = { extras = ["all"], version = "^0.104.1" }
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
fastapi-another-jwt-auth = "^0.1.10"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
autoflake = "^2.2.1"
flake8 = "^6.1.0"
isort = "^5.13.0"
pre-commit = "^3.5.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.flake8]
max-line-length = 120
per-file-ignores = ["src/models/__init__.py:F401"]

[tool.black]
line-length = 120
target-version = ["py310"]
# isort
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.isort]
profile = "black"

[tool.mypy]
# ignore_missing_imports = true
exclude = [
    ".idea/",
    '.venv/',
    'vscode/*',
]
